With last will, the server knows when the device is offline

+++ server/mqtt/device/management/commands/monitor_devices.py
@@ -1,5 +1,4 @@
 import datetime
-import importlib
 import sys
 
 from django.core.management.base import BaseCommand
@@ -9,23 +8,26 @@ from django.utils.module_loading import import_string
 import paho.mqtt.client as mqtt
 
 from device.models import Device, Operation, Arg
-from device.utils import get_id_from_spec_topic
+from device.utils import match_topic, raise_if
 
 
-# TODO: change so that last will message will be parsed
-def on_message(client, userdata, msg):
-    device_id = get_id_from_spec_topic(msg.topic)
-    device, created = Device.objects.create_or_update(
+def online(msg, device_id):
+    # TODO: change so that if a device is already seen as online,
+    #       we should somehow treat that
+    #       this may happen when the keepalive time of a device is long enough
+    #       such that it is unresponsive but the broker has not yet figured
+    #       that out because the keepalive has not expired yet
+    device, created = Device.objects.update_or_create(
         device_id=device_id, defaults={'online': True}
     )
 
     if created:
         sys.stdout.write('Added device with id: {}\n'.format(device_id))
     else:
-        device.operations.filter(
-            created_at__lt=device.last_offline).delete()
+        if device.last_offline:
+            device.operations.filter(
+                created_at__lt=device.last_offline).delete()
 
-    # TODO: change online state of device somehow
     parser = import_string(settings.SPEC_PARSER)
     validated_data = parser(msg.payload)
     validated_data['device_id'] = device.id
@@ -40,10 +42,44 @@ def on_message(client, userdata, msg):
         Arg.objects.create(**arg)
 
 
+def offline(msg, device_id):
+    try:
+        device = Device.objects.get(device_id=device_id)
+    except Device.DoesNotExist:
+        pass
+
+    device.online = False
+    device.last_offline = datetime.datetime.now()
+    device.save()
+
+    sys.stdout.write('Device {} is offline\n'.format('device_id'))
+
+
+def on_message(client, userdata, msg):
+    is_connect, connect_args = match_topic(
+        settings.MQTT_DEVICE_SPEC_TOPIC, msg.topic)
+    is_disconnect, disconnect_args = match_topic(
+        settings.MQTT_DEVICE_OFFLINE, msg.topic)
+
+    if is_connect:
+        online(msg, connect_args[0])
+    elif is_disconnect:
+        offline(msg, disconnect_args[0])
+
+
 def on_connect(client, userdata, flags, rc):
     sys.stdout.write('Connected to broker\n')
     sys.stdout.write('Waiting for incoming messages ...\n')
+
+    raise_if(
+        settings.MQTT_DEVICE_SPEC_TOPIC.count('+') != 1,
+        'Spec topic should have a single parameter, device id (one +)')
+    raise_if(
+        settings.MQTT_DEVICE_OFFLINE.count('+') != 1,
+        'Offline topic should have a single parameter, device id (one +)')
+
     client.subscribe(settings.MQTT_DEVICE_SPEC_TOPIC)
+    client.subscribe(settings.MQTT_DEVICE_OFFLINE)
 
 
 class Command(BaseCommand):
